requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.10
beautifulsoup4==4.12.3
